Get status, LocalAssetsManager is one class again
